requirements.txt
praw
playwright
google-generativeai
streamlit
